--- a/src/main/java/org/brapi/test/BrAPITestServer/repository/BrAPIRepository.java
+++ b/src/main/java/org/brapi/test/BrAPITestServer/repository/BrAPIRepository.java
@@ -3,8 +3,8 @@ package org.brapi.test.BrAPITestServer.repository;
 import java.io.Serializable;
 import java.util.List;
 import java.util.Optional;
-import java.util.UUID;
 
+import org.brapi.test.BrAPITestServer.exceptions.InvalidPagingException;
 import org.brapi.test.BrAPITestServer.model.entity.BrAPIPrimaryEntity;
 import org.brapi.test.BrAPITestServer.service.SearchQueryBuilder;
 import org.springframework.data.domain.Page;
@@ -15,9 +15,9 @@ import org.springframework.data.repository.NoRepositoryBean;
 @NoRepositoryBean
 public interface BrAPIRepository<T extends BrAPIPrimaryEntity, ID extends Serializable> extends JpaRepository<T, ID> {
 
-	public Page<T> findAllBySearchAndPaginate(SearchQueryBuilder<T> searchQuery, Pageable pageReq);
+	public Page<T> findAllBySearchAndPaginate(SearchQueryBuilder<T> searchQuery, Pageable pageReq) throws InvalidPagingException;
 
-	public Page<T> findAllBySearchPaginatingWithFetches(SearchQueryBuilder<T> searchQuery, Pageable pageReq);
+	public Page<T> findAllBySearchPaginatingWithFetches(SearchQueryBuilder<T> searchQuery, Pageable pageReq) throws InvalidPagingException;
 
 	public List<T> findAllBySearch(SearchQueryBuilder<T> searchQuery);
 
@@ -29,5 +29,5 @@ public interface BrAPIRepository<T extends BrAPIPrimaryEntity, ID extends Serial
 	
 	public <S extends T> void refresh(S entity);
 
-	public void fetchXrefs(Page<T> page, Class<T> searchClass);
+	public void fetchXrefs(Page<T> page, Class<T> searchClass) throws InvalidPagingException;
 }