--- a/src/main/resources/application.properties.template
+++ b/src/main/resources/application.properties.template
@@ -22,3 +22,12 @@ spring.mvc.dispatch-options-request=true
 security.oidc_discovery_url=https://example.com/auth/.well-known/openid-configuration
 security.enabled=true
 security.issuer_url=http://example.com/issuerurl
+
+# This should either be set in accordance with a maximum number of SQL parameters (on JOIN FETCHES of collections,
+# if there is more than one collection the IDs of each entity need to be passed through as parameters, and there is a SQL
+# maximum of 65535.  See GermplasmService.findGermplasmEntities()),
+# whatever returns in a reasonable amount of time,
+# or if you want to limit for the sake of server efficiency.
+paging.page-size.max-allowed=65000
+
+paging.page-size.default=1000