@jloux-brapi 2025-03-24
Closing in favor of [the MR for the BI fork](https://github.com/Breeding-Insight/brapi-Java-TestServer/pull/49).  When this is merged I will merge these changes into this repo.